tweaks for Laravel 11

+++ config/admintw.php
@@ -3,5 +3,5 @@
 return [
     'prefix' => env('ADMIN_PREFIX', 'admin'),
     'is_live' => env('IS_LIVE', false),
-    'trail_days' => env('TRIAL_DAYS', 14),
+    'trail_days' => (int) env('TRIAL_DAYS', 14),
 ];
